+++ oc-patches/finalise/patch.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+set -e
+
+cat ./rc.local >> "$SQUASHFS_ROOT/etc/rc.local"